Refactor MistweaverMonk.lua for performance and readability
I have refactored the `MistweaverMonk.lua` script to improve performance, readability, and maintainability.

The main changes are:

- Introduced a new `UnitScanner` module that iterates over units only once per frame, significantly improving performance.
- Moved large data tables (`dispelList`, `sootheList`, `debuffList`) to a separate `MonkData.lua` file to improve code organization.
- Refactored the APLs to use the new `UnitScanner` module, making them more concise and easier to understand.
- Replaced global slash commands with `Bastion.Command` to avoid polluting the global namespace.
- Removed redundant helper functions and custom units.
